// src/plugins/twitch/types/twitch-tv-get-users-response.ts
export interface TwitchTvGetUsersResponse {
  data: {
    broadcaster_type: string;
    description: string;
    display_name: string;
    email: string;
    id: string;
    login: string;
    offline_image_url: string;
    profile_image_url: string;
    type: string;
    view_count: number;
  }[];
}